Added Modifications to the API calls as requested by the frontend team. Also added validation util methods.Strategy change of blobs creation in S3

+++ src/main/java/com/software/upskilled/repository/CourseMaterialRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CourseMaterialRepository extends JpaRepository<CourseMaterial, Integer>
+public interface CourseMaterialRepository extends JpaRepository<CourseMaterial, Long>
 {
     CourseMaterial findByTitle( String courseTitle );
 }
